requirements.txt
# Core dependencies
PyYAML>=6.0
Pillow>=9.0.0
requests>=2.28.0

# RSS feed parsing
feedparser>=6.0.10

# Retry logic
tenacity>=8.2.0

# E-paper display hardware support
spidev>=3.5
gpiozero>=2.0.0
colorzero>=2.0

# Date/time utilities
python-dateutil>=2.8.0
